requirements.txt
"flask<3"
"werkzeug<3"
flask_login
flask_sqlalchemy
flask_migrate
fpdf <> flask_weasyprint
openpyxl
python-dotenv-1.0.1